Bumping version to 2 in setup.py. Moving to src directory.
