+++ .gitignore
@@ -1,4 +1,5 @@
 build/
+dist/
 __pycache__
 monerorequest.egg-info
 htmlcov
