--- monerorequest/__init__.py
@@ -1,1 +1,0 @@
-from .monerorequest import *  # noqa: F403
