+++ setup.py
@@ -2,11 +2,11 @@ from setuptools import setup
 
 setup(
     name='monerorequest',
-    version='1.2.0',
+    version='2.0.0',
     author="Luke Profits",
     description="monerorequest is an easy way to create/decode Monero Payment Requests.",
     url="https://github.com/lukeprofits/monerorequest",
-    packages=['monerorequest'],
+    packages=['src/monerorequest'],
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
